requirements.txt
# requirements.txt
openai>=0.27.0
python-dotenv
faiss-cpu>=1.7.4   # optional but recommended; on some platforms use 'faiss-gpu' if available
tqdm
numpy
PyPDF2
regex
scikit-learn
